Move StressTester class to utils file
